Semester_2/DE/telegram_bot/requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25
aiohttp==3.9.1
pandas==2.2.0
openpyxl==3.1.2
yadisk==3.2.0 